Dev: get rid of token and cctp gas estimation utils
